change float

+++ src/PwmSin.c
@@ -7,14 +7,14 @@ void PwmSin_Gen(uint8_t *array, uint16_t size)
 {
 	uint16_t i; // loop counter
 	uint8_t j;
-	double try;
-	double sin_var;
-	double sample;
-	double step_value;
+	float try;
+	float sin_var;
+	float sample;
+	float step_value;
 	uint16_t offset;
 
 	sample = 0.0;
-	step_value = (1.0 / (double)size);
+	step_value = (1.0 / (float)size);
 	offset = (size / BIT_PER_BYTE);
 
 	for (i = 0; i < size; i++)
